Refactor dts-process.js and add dtsProcessEventRule.json
