+++ lib/functions/dts-process.js
@@ -3,10 +3,9 @@ const logger = require('../helpers/logging')
 const pg = require('../helpers/db')
 const invokeLambda = require('../helpers/invoke-lambda')
 const Joi = require('joi')
-const axios = require('axios')
-const { HTTP_NOT_FOUND } = require('../constants')
 const deleteStation = require('../helpers/imtd-api').deleteStation
 const getRloiIds = require('../helpers/imtd-api').getRloiIds
+const getImtdApiResponse = require('../helpers/imtd-api').getImtdApiResponse
 
 async function insertStation (stationDataArray) {
   try {
@@ -21,21 +20,6 @@ async function insertStation (stationDataArray) {
   } catch (error) {
     logger.error('Database error processing stationData', error)
     throw error
-  }
-}
-
-async function getImtdApiResponse (stationId) {
-  const hostname = 'imfs-prd1-thresholds-api.azurewebsites.net'
-  try {
-    return await axios.get(`https://${hostname}/Location/${stationId}?version=2`)
-  } catch (error) {
-    if (error.response?.status === HTTP_NOT_FOUND) {
-      logger.info(`Station ${stationId} not found (HTTP Status: 404)`)
-    } else {
-      const message = error.response?.status ? `HTTP Status: ${error.response.status}` : `Error: ${error.message}`
-      throw Error(`IMTD API request for station ${stationId} failed (${message})`)
-    }
-    return {}
   }
 }
