+++ lib/functions/imtd-process.js
@@ -1,12 +1,11 @@
 const parseThresholds = require('../models/parse-thresholds')
-const axios = require('axios')
 const logger = require('../helpers/logging')
 const pg = require('../helpers/db')
 const invokeLambda = require('../helpers/invoke-lambda')
 const deleteThresholds = require('../helpers/imtd-api').deleteStation
 const getRloiIds = require('../helpers/imtd-api').getRloiIds
 const tableName = 'station_imtd_threshold'
-const { HTTP_NOT_FOUND } = require('../constants')
+const getImtdApiResponse = require('../helpers/imtd-api').getImtdApiResponse
 
 async function insertThresholds (stationId, thresholds) {
   try {
@@ -28,21 +27,6 @@ async function insertThresholds (stationId, thresholds) {
   } catch (error) {
     logger.error(`Database error processing thresholds for station ${stationId}`, error)
     throw error
-  }
-}
-
-async function getImtdApiResponse (stationId) {
-  const hostname = 'imfs-prd1-thresholds-api.azurewebsites.net'
-  try {
-    return await axios.get(`https://${hostname}/Location/${stationId}?version=2`)
-  } catch (error) {
-    if (error.response?.status === HTTP_NOT_FOUND) {
-      logger.info(`Station ${stationId} not found (HTTP Status: 404)`)
-    } else {
-      const message = error.response?.status ? `HTTP Status: ${error.response.status}` : `Error: ${error.message}`
-      throw Error(`IMTD API request for station ${stationId} failed (${message})`)
-    }
-    return {}
   }
 }
