+++ lib/helpers/imtd-api.js
@@ -1,4 +1,7 @@
 const pg = require('./db')
+const axios = require('axios')
+const { HTTP_NOT_FOUND } = require('../constants')
+const logger = require('./logging')
 
 async function deleteStation (stationId, tableName) {
   await pg(tableName).where({ station_id: stationId }).delete()
@@ -18,4 +21,19 @@ async function getRloiIds ({ limit, offset } = {}) {
   }
 }
 
-module.exports = { deleteStation, getRloiIds }
+async function getImtdApiResponse (stationId) {
+  const hostname = 'imfs-prd1-thresholds-api.azurewebsites.net'
+  try {
+    return await axios.get(`https://${hostname}/Location/${stationId}?version=2`)
+  } catch (error) {
+    if (error.response?.status === HTTP_NOT_FOUND) {
+      logger.info(`Station ${stationId} not found (HTTP Status: 404)`)
+    } else {
+      const message = error.response?.status ? `HTTP Status: ${error.response.status}` : `Error: ${error.message}`
+      throw Error(`IMTD API request for station ${stationId} failed (${message})`)
+    }
+    return {}
+  }
+}
+
+module.exports = { deleteStation, getRloiIds, getImtdApiResponse }
